🚚 Separate server from index file in backend
